Sync'd to SMF r10396  & Fixed a couple of languages strings. [Bug 4566] (Help.english.php, ManagePermissions.english.php)  ! Fixed validation error and header bar markup. [Bug 4549] (ModerationCenter.template.php) (Note from commit: We have several cases in admin where HTML validation is borked if there are no items displayed in a table. The problem is that the php was written without taking into account the possibility of empty tables. I'll fix them as I come across them but wont go looking for them at this stage.) (NB, I fixed this differently to them, i.e. not having an extra if after the foreach)  * Fixed newsletter IE7 issue and some other IE issues. (ManageNews.template.php, Login.template.php, index.css, Warm/index.css, rtl.css, Warm/rtl.css, ie6.css, ie7.css, Warm/ie6.css, Warm/ie7.css) [Bug 4567] (NB: These don't appear to make a difference to me, but might on IE, haven't tested them)  * Incorrect id's fixed. (SplitTopics.template.php) [Bug 4582]  * Minor markup fixes. (Profile.template.php) [Bug 4607] (One did not apply to us!)  ! Fixed display of spider log title/information. (ManageSearchEngines.php) [Bug 4608]  ! Broken function call in ManageMail (ManageMail.php)  ! Credits updates. (Credits.php, Who.english.php) (NB: I've added Brett as consulting dev since some of his work has gone into the patches here, and added the server team but not any of the miscellaneous people moves since they're not part of Team Wedge)  * Added array flip for buttonlists to fix untracked ltr/rtl issues in IE6/7. (rtl.css, index.template.php)  ! messageIconChecks_disable should be enable [Bug 4095] (Display.php, MessageIndex.php, Search2.php, Unread.php, Unreadreplies.php, SSI.php) [2.1 branch but was in the tracker for ages so fair game]  ! Multitable deletes possible in MySQL. (ManageBans.php, ManageMaintenance.php, ManagePermissions.php) [2.1 branch, ignored the == to === micro optimisations]  ! Checking for search API files could cause an open_basedir error. (ManageSearch.php) [Bug 4251]  ! Visual verification could cause undefined index errors in some cases. (Subs-Editor.php) [Bug 4546]  ! add_index function failed on duplicate indexes. (Class-DBPackages.php) [Bug 4536]  ! Allow search methods to modify search settings. (ManageSearch.php) [Bug 4606]
Not added from SMF SVN
 ! revs 10347-10350 (branches in SVN, plus backporting a change to 1.1.x branch)
 ! rev 10351 (SSI change to prevent functions being called from GET when as a guest; since protection is added at function level rather than the file level, we don't need to add this)
 ! revs 10352-10355 (SSI changes from 2.0 being backported to 1.0 and 1.1, plus version number bumps for these versions)
 ! rev 10356 (typo fix in 1.1.x SSI)
 ! revs 10360, 10363 and 10368 in relation to admin submenu not having clear on narrower widths properly... cannot reproduce this myself though their fixes seem to be generic to IE6 too (and revolve around putting a br clear in the generic template. Also, they add a white-space: pre, we already have a related white-space: nowrap in our equivalent menu CSS)
 ! rev 10361 (removes the 3px left padding in h3.titlebg only in admin panel)
 ! rev 10362 (missing class="icon" on some h3s in Admin and Themes templates... we don't use that any more)
 ! rev 10370 (?fin20 to ?rc5 for URLs)
 ! rev 10371 (bump to 2.0 RC5 - save that for another commit I think)
 ! rev 10373 (creation of 2.1 branch in SVN)
 ! rev 10374 (1.1.x ManageNews version change)
 ! rev 10375 (removal of an extra space in 1.1.x SSI)
 ! revs 10376-10379, 10381-10383 (preparation of patch files for just the security changes, 1.1 patch, 2.0 patch, patch cleanup, 1.1 patch, 2.0 patch, version bumps in installer/upgrader)
 ! rev 10380 (Admin language file in 1.1.x cleanup)
 ! rev 10384 (1.0/1.1 changelog notices for security)
 ! rev 10385 ([2.1 branch] check for @ before attempting to query supplied username in email field)
 ! rev 10387 ([2.1 branch] replace all calls to time() with $_SERVER['REQUEST_TIME']... aside from a (IMO) negligable benefit, that relies on the server setting it up, and I don't expect it to be completely consistent on all environments)
 ! rev 10389 (tweaks for manual/wiki changes)
 ! rev 10390 (help section not right in Core)
 ! rev 10396 ([2.1 branch] rewrites of version_compare to use "nice syntax" with symbols - we don't have any outside the installer to worry about, and apparently there's a search bug fixed but I don't see what it is from the changelog)

+++ Sources/ManagePermissions.php
@@ -1043,22 +1043,13 @@ function GeneralPermissionSettings($return_config = false)
 
 		// Clear all deny permissions...if we want that.
 		if (empty($modSettings['permission_enable_deny']))
-		{
 			wesql::query('
-				DELETE FROM {db_prefix}permissions
+				DELETE FROM {db_prefix}permissions, {db_prefix}board_permissions
 				WHERE add_deny = {int:denied}',
 				array(
 					'denied' => 0,
 				)
 			);
-			wesql::query('
-				DELETE FROM {db_prefix}board_permissions
-				WHERE add_deny = {int:denied}',
-				array(
-					'denied' => 0,
-				)
-			);
-		}
 
 		// Make sure there are no postgroup based permissions left.
 		if (empty($modSettings['permission_enable_postgroups']))
